requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.202
botocore==1.12.202
bpemb==0.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorcet==2.0.1
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
gensim==3.8.0
gevent==1.4.0
greenlet==0.4.15
idna==2.8
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.0
pandocfilters==1.4.2
param==1.9.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyct==0.4.6
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
runway-python==0.4.0
s3transfer==0.2.1
scipy==1.3.0
Send2Trash==1.5.0
sentencepiece==0.1.82
six==1.12.0
smart-open==1.8.4
terminado==0.8.2
testpath==0.4.2
torch==1.1.0.post2
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wget==3.2
widgetsnbextension==3.5.1
